Day-9 Flex Box
On Day 9 of our CSS learning journey, we delve into the powerful Flexbox layout module. Flexbox, short for "Flexible Box", is a CSS3 layout mode designed to provide a more efficient way to arrange and align elements within a container, even when their size is unknown and/or dynamic. This layout model is particularly useful for creating responsive web designs.